Added baseKeyPair to dedup common behaviour, other minor PR changes.
Ticket: BGX-44
