+++ src/coin/xtz/keyPair.ts
@@ -1,7 +1,8 @@
 import * as crypto from 'crypto';
-import { DefaultKeys, ExtendedKeys } from '../baseCoin/iface';
+import { DefaultKeys } from '../baseCoin/iface';
 import * as CryptoUtils from '../../utils/crypto';
 import { isPrivateKey, isPublicKey, isSeed, KeyPairOptions } from '../baseCoin/iface';
+import { BaseKeyPair } from '../baseCoin/baseKeyPair';
 import * as Utils from './utils';
 import blake2b from 'blake2b';
 import { HDNode, ECPair } from 'bitgo-utxo-lib';
@@ -11,19 +12,15 @@ const DEFAULT_SEED_SIZE_BYTES = 16;
 /**
  * Tezos keys and address management.
  */
-export class KeyPair {
-  // Implementation of the HD protocol (BIP32). Only available when creating a KeyPair from a seed,
-  // or extended keys
-  private hdNode?: HDNode;
-  private keyPair: ECPair;
-
+export class KeyPair extends BaseKeyPair {
   /**
    * Public constructor. By default, creates a key pair with a random master seed.
    *
-   * @param source Either a master seed, a private key (extended or raw), or a public key
+   * @param {KeyPairOptions} source Either a master seed, a private key (extended or raw), or a public key
    *     (extended, compressed, or uncompressed)
    */
   constructor(source?: KeyPairOptions) {
+    super(source);
     if (!source) {
       const seed = crypto.randomBytes(DEFAULT_SEED_SIZE_BYTES);
       this.hdNode = HDNode.fromSeedBuffer(seed);
@@ -44,10 +41,11 @@ export class KeyPair {
 
   /**
    * Build a Hierarchical Deterministic node or an ECPair from a private key.
+   * Specific Tezos implementation
    *
-   * @param prv An extended or raw private key
+   * @param {string} prv An extended or raw private key
    */
-  private recordKeysFromPrivateKey(prv: string): void {
+  protected recordKeysFromPrivateKey(prv: string): void {
     if (CryptoUtils.isValidXprv(prv)) {
       this.hdNode = HDNode.fromBase58(prv);
     } else if (CryptoUtils.isValidPrv(prv)) {
@@ -62,10 +60,11 @@ export class KeyPair {
 
   /**
    * Build a Hierarchical Deterministic node or an ECPair from a public key.
+   * Specific Tezos implementation
    *
    * @param {string} pub - An extended, compressed, or uncompressed public key
    */
-  private recordKeysFromPublicKey(pub: string): void {
+  protected recordKeysFromPublicKey(pub: string): void {
     if (CryptoUtils.isValidXpub(pub)) {
       this.hdNode = HDNode.fromBase58(pub);
     } else if (CryptoUtils.isValidPub(pub)) {
@@ -81,7 +80,7 @@ export class KeyPair {
   /**
    * Return Tezos default keys with the respective prefixes
    *
-   * @returns The keys in the protocol default key format
+   * @returns {DefaultKeys} The keys in the protocol default key format
    */
   getKeys(): DefaultKeys {
     // Always use the compressed version to be consistent
@@ -99,25 +98,9 @@ export class KeyPair {
   }
 
   /**
-   * Get the extended public key, and the private key if one is available. This is only possible
-   * when the key pair was created from a seed or extended keys.
-   */
-  getExtendedKeys(): ExtendedKeys {
-    if (!this.hdNode) {
-      throw new Error('Cannot get extended keys');
-    }
-    const result: ExtendedKeys = {
-      xpub: this.hdNode.neutered().toBase58(),
-    };
-    // A neutered HD node means it only contains the public key information
-    if (!this.hdNode.isNeutered()) {
-      result.xprv = this.hdNode.toBase58();
-    }
-    return result;
-  }
-
-  /**
    * Get a public address.
+   *
+   * @returns {string} The public address
    */
   getAddress(): string {
     const pub = this.keyPair.Q.getEncoded(true);
